requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
django-cors-headers==4.4.0
django-filter==24.3
Pillow==10.4.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
python-dotenv==1.0.1